+++ src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './Encryptor';
 export * from './ExceptionFilter';
+export * from './QueryBuilder';
 export * from './ResponseInterceptor';